[FIX] coop_shift: tests
