+++ coop_shift/tests/__init__.py
@@ -1,2 +1,1 @@
-from . import common
 from . import test_coop_shift
